Correct default batch job output file name
In version 14.03.2 was using "slurm_<jobid>_4294967294.out" due to
error in job array logic.

+++ src/slurmd/slurmstepd/slurmstepd_job.c
@@ -148,7 +148,7 @@ static char *
 _batchfilename(stepd_step_rec_t *job, const char *name)
 {
 	if (name == NULL) {
-		if (job->array_task_id == (uint16_t) NO_VAL)
+		if (job->array_task_id == NO_VAL)
 			return fname_create(job, "slurm-%J.out", 0);
 		else
 			return fname_create(job, "slurm-%A_%a.out", 0);
